requirements.txt

flask
python-dotenv
werkzeug